Split JobArchiveRequest from other type definitions into worker API package and move remaining types to separate package
