+++ archiveworker/type/job_status.py
@@ -1,0 +1,31 @@
+# Moodle Quiz Archive Worker
+# Copyright (C) 2025 Niels Gandraß <niels@gandrass.de>
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+from enum import StrEnum
+
+
+class JobStatus(StrEnum):
+    """
+    Status values a single quiz archive worker job can have
+    """
+    UNINITIALIZED = 'UNINITIALIZED'
+    AWAITING_PROCESSING = 'AWAITING_PROCESSING'
+    RUNNING = 'RUNNING'
+    WAITING_FOR_BACKUP = 'WAITING_FOR_BACKUP'
+    FINALIZING = 'FINALIZING'
+    FINISHED = 'FINISHED'
+    FAILED = 'FAILED'
+    TIMEOUT = 'TIMEOUT'
